CCompiler: Corrected empty lines in the dispatch table.
The selector had been "nil", which is zero, which is a valid selector, so there
was a theoretical possibility of dispatching to a wrong method when the object
didn't understand a particular selector.  But in reality, selector zero would
always be chosen for a method in Object (eg. currently it's chosen for "!="),
so the problem would never occur.  Fixed it anyway.