Fixed a crash when running under Jolt.
The parser was giving some function names as Symbols, which crashed the
selector-info stuff when the symbol was used as a key for a dictionary.  I'm
not sure why that wasn't a problem earlier; it seems like at: failed but
at:put: didn't somehow.  Anyhow, we now make sure the keys are turned into
strings.