CCompiler: Removed the "fastcall" attribute from Dispatch_().
At first I added a "defined(__i386__)" condition to the #ifdef, in order to
shut up the warnings when compiling on the Mac.  Then I noticed that it was
actually a little faster today without the "fastcall" (on x86).  So I just
removed it entirely.  (Note, however, that I wouldn't have checked it in in the
first place if it didn't help on the first day I tried it...)