Added "shared-ints" option.
Previously, all Int (and Char) literals were defined local (but static) to each
method.  Turning on the 'shared-ints' build setting now allocates a global pool
of ints.  Even in tests/syntax, this saves about 3k (almost 1%) of the total
program size (with debugging on).  However, it is not currently turned on by
default.