X86Compiler: Fixing stack alignment under Mac OSX.
The Mac OSX ABI requires that the stack be aligned to a 16-byte boundary at
every function call.  We can get away with violating that for some of our own
calls (eg. to Dispatch_() during a send), but not when calling anything that
might result in a call to the OS libraries.  This makes codegen considerably
trickier, and X86 machine code becomes less like a funny sort of bytecode.