X86Compiler: First successful creation of a new class.
The main thing that kept this from working was that the new classes were
getting GC'd.  The only reference was from within the "returner" method, which
is not traced by the GC.  (Another important bug was that the new class's
functions were being built into "Standard CImplementation Class" instead of the
new class.)

Also improved the creating of the ClassInfo, adding "addedFields" and using
Symbols instead of Strings.