Tagged Ints.
Ints that fit into 31 bits (and sometimes into 63 bits on 64-bit systems) are
now tagged, which can prevent a lot of consing for some programs.  Trylon never
bothered with this before, as the theory is that most programs of interest to
Trylon's author deal with strings much more often than numbers, so it wasn't
worth the expense, however small, at dispatch time to check for the tag.  But a
lot of Ints were being consed both in the "bench" program, and when building
the dispatch table in the compiler.  So here are tagged Ints, which do make a
big difference in "bench" at least.