Class-based scripting is now the default.
If you call Trylon with a filename, including by running a file starting with
"#!/usr/bin/env trylon", it will now try to load that file as a class and run
its "main:" method (passing the arguments).  This is only supported on X86
systems.  If you have old-style scripts (which treat the file as a sequence of
statements, rather than the definition of a class), change your shebang to
"#!/usr/bin/env trylon-interpreter".