Improved ExistingProto.
'lookup-instance-function:' had still been checking the case of the function
name to determine whether it was referring to a subproto.  That was merely a
heuristic, not reliable.  We've had the subproto info in the class object for a
while now (as long as 'support-subprotos' is on); 'function-for:in:' had
already been using that.  So 'lookup-instance-function:' now uses it too.