Reinstated the "uses" directive.
This existed in Trylon 1, but I think it fell out in Trylon 2.  So far, the
only use of "uses" has been when Main uses Standard, which was set up by the
compiler rather than in Main's Trylon code.  So nobody noticed its absence.

This included a fix for using a CompiledFunction in code compiled by the
X86Compiler.

I notice that in C compiled code, a function from a "used" class is visible
outside the "using" class, but in X86 compiled code, it is available only
inside the "using" class.  That is, in tests/syntax, you can call "UsedClass
value" from C-compiled code but not in a script.